requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
fastapi==0.112.1
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
wsproto==1.2.0
